Rename "meshes" code and add wireframe style
Soon, we'll add the volume style
